oriente-backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Email validation
email-validator==2.1.0

# Utilitários
python-multipart==0.0.6
python-dotenv==1.0.0
